+++ src/main/java/com/devagit/springbootstudy/repository/heart/HeartJpaRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HeartJpaRepository extends JpaRepository<Heart, Integer> {
 
-    Heart findByPostIdAndUserId(int postId, String userId);
+    Heart findByPostIdAndUserId(long postId, String userId);
 
-    void deleteByPostIdAndUserId(int postId, String userId);
+    void deleteByPostIdAndUserId(long postId, String userId);
 
-    int countByPostId(int postId);
+    int countByPostId(long postId);
 }
