+++ src/main/java/com/devagit/springbootstudy/request/post/UpdateRequest.java
@@ -9,15 +9,15 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Getter
 public class UpdateRequest {
-    private int id;
-    private int subCategoryId;
+    private long id;
+    private  int subCategoryId;
     private String userId;
     private String title;
     private String contents;
     private String source;
     private LocalDateTime createdAt;
 
-    public UpdateRequest(int id, int subCategoryId, String userId, String title, String contents, String source, LocalDateTime createdAt) {
+    public UpdateRequest(long id,int subCategoryId, String userId, String title, String contents, String source, LocalDateTime createdAt) {
         this.id = id;
         this.subCategoryId = subCategoryId;
         this.userId = userId;
